Oddzieliłem UserPanelModal z glownego kodu  	modified:   src/App.jsx  	modified:   src/components/UserPanel/UserPanel.jsx  	modified:   src/components/UserPanel/UserPanel.module.css  	new file:   src/components/UserPanel/UserPanelModal.jsx  	new file:   src/components/UserPanel/UserPanelModal.module.css
