+++ src/App.jsx
@@ -12,8 +12,8 @@ import HomeTab from './components/HomeTab/HomeTab.jsx';
 import CurrencyTable from './components/CurrencyTable/CurrencyTable.jsx';
 
 const Login = lazy(() => import('./pages/LoginPage/LoginPage.jsx'));
-const Register = lazy(() => import('./pages/RegisterPage/RegisterPage.jsx'));  
-// const Register = lazy(() => import('./pages/Dashboard/Dashboard.jsx'));  
+// const Register = lazy(() => import('./pages/RegisterPage/RegisterPage.jsx'));  
+const Register = lazy(() => import('./components/UserPanel/UserPanel.jsx'))
 const Dashboard = lazy(() => import('./pages/Dashboard/Dashboard.jsx'));
 const NotFound = lazy(() => import('./pages/NotFound.jsx'));
 const VerifyEmail = lazy(() => import('./pages/VerifyEmail.jsx'));
