+++ src/components/UserPanel/UserPanel.module.css
@@ -93,5 +93,5 @@
   .container_input {
     display: flex;
     justify-content: center;
-    margin-top: 20px;
+    margin: 30px 40px;
   }
